Turtle

Description

This module provides a domain specific language for moving a turtle on the screen. The idea is that a program is a list of instructions (for instance GoLeft 300). Programs can then be sequenced and combined.

Documentation

The turtle type(s)

data Operation

Change from hand in #1: We have removed a couple of operations, instead using for instance Forward (-1) as our old Backward 1

Constructors

GoLeft Double

Forward Double

PenUp

PenDown

SetColor TColor

ShowTurtle

HideTurtle

End

type Actions = [Operation]

To simplify for the user and to make parallel composition easier to use, we decided to split our previous Program type (from hand in #1) into two types. Actions is what we previously defined as a Program, a list of Operations. Now a Program is a list of a list of Operations, this way the user doesn't need to give an explicit list when writing a program

type Program = [Actions]

data Turtle

Data type for keepig the state of a turtle

Constructors

Turtle

Fields

position :: TPoint

degree :: Double

isPenDown :: Bool

turColor :: TColor

isVisible :: Bool

isDead :: Bool

data TColor

Modeling Colors

Constructors

Yellow

Blue

Red

Green

type TPoint = (Double, Double)

A simple cartesian plane point

Primitive operations

forward :: Int -> Program

Primitive operations Forward replicates itself to make our graphical animation look smoother

backward :: Int -> Program

left :: Double -> Program

color :: TColor -> Program

penup :: Program

pendown :: Program

die :: Program

tforever :: Program -> Program

Given a program, repeat it in infinity, users will then exploit haskell's lazy properties to get information from it. Works on parallel composition as well.

times :: Int -> Program -> Program

times takes an existing program, and a number of times to repeat the program. Then it returns a new program repeated that number of times. Works on parllel composition as well.

showturtle :: Program

hideturtle :: Program

Derived operations

right :: Double -> Program

Derived operation

lifespan :: Int -> Program -> Program

Potentially shorten a program to be limited by a given amount of steps. Works on parallel composition as well

Constructor functions

identityTurtle :: Turtle

Constructors Create a standard default turtle

Combinator functions

(<*>) :: Program -> Program -> Program

Combinators for all programs in the list of programs to run, add the given program at the end

(<|>) :: Program -> Program -> Program

Parallel composition Adds a program to the list of parallel programs

Run function

run :: Turtle -> Operation -> Turtle

Abstract run function that takes a turtle and an operation and gives a new turtle back. Will be used by other concrete run funcs This is where most of the work happens because we have a deep embedding.

runTextual :: Turtle -> Program -> IO ()

Run function for printing a program to the terminal, can print both finite and infinite programs.