TurtleGraphics

Description

A graphical run function for the turtle DSL

Documentation

runGraphical :: Turtle -> Program -> IO ()

Creates a GTK window, creates two state variables (drawState,history) drawState keeps track of what to draw in each round and history keeps track of everthing that has previously been drawed to screen.

We're using the timeoutAdd function to define a timer that calls exposeEvent on the window. This is why we need the two state variables, since we are calling exposeEvent on small chunks of the total data to draw it's not possible to call a function that recurses to itself and increases some internal state that would cause GTK to freeze